swww don't refresh if steam is running

+++ hyprland/.config/hypr/swww_randomize.sh
@@ -13,23 +13,26 @@ fi
 INTERVAL="$2"
 
 while true; do
-	find "$1" \
-		| while read -r img; do
-			echo "$((RANDOM % 1000)):$img"
-	done \
-		| sort -n | cut -d':' -f2- \
-		| while read -r img; do
-
-			killall waybar &
-			
-			magick "$img" -gravity Center -extent 1.005:1 ~/.cache/rofi.bmp &
-
-			swww img "$img" &
-
-			wallust run "$img" --quiet -d ~/.config/wallust/
-
-			waybar &
-
-			sleep $INTERVAL
-		done
-	done
+	if ! pgrep -x "steam" > /dev/null
+	then
+		find "$1" \
+			| while read -r img; do
+				echo "$((RANDOM % 1000)):$img"
+		done \
+			| sort -n | cut -d':' -f2- \
+			| while read -r img; do
+	
+				killall waybar &
+				
+				magick "$img" -gravity Center -extent 1.005:1 ~/.cache/rofi.bmp &
+	
+				swww img "$img" &
+	
+				wallust run "$img" --quiet -d ~/.config/wallust/
+	
+				waybar &
+	
+				sleep $INTERVAL
+			done
+	fi
+done
